Code reorganization.
