+++ package.json
@@ -78,9 +78,12 @@
     "mocha": "^5.2.0",
     "postcss-loader": "^3.0.0",
     "postcss-nesting": "^7.0.0",
+    "prop-types": "^15.7.2",
     "react": "^16.8.1",
     "react-dom": "^16.8.1",
+    "react-redux": "^7.2.0",
     "react-rnd": "^10.0.0",
+    "redux": "^4.0.5",
     "sinon": "^7.2.3",
     "style-loader": "^0.23.1",
     "webpack": "^4.29.3",
