+++ src/editorobserver.js
